+++ pkg/analyzer/benchmark/errors_in_all_libraries.dart
@@ -34,7 +34,6 @@ void main(List<String> args) {
     AnalysisEngine.instance.clearCaches();
 
     AnalysisOptionsImpl options = new AnalysisOptionsImpl();
-    options.strongMode = true;
     options.strongModeHints = true;
 
     PhysicalResourceProvider resourceProvider =