+++ pkg/analyzer/test/src/summary/resynthesize_ast_test.dart
@@ -32,7 +32,6 @@ import 'summary_common.dart';
 
 main() {
   defineReflectiveSuite(() {
-    defineReflectiveTests(ResynthesizeAstSpecTest);
     defineReflectiveTests(ResynthesizeAstStrongTest);
     defineReflectiveTests(ApplyCheckElementTextReplacements);
   });
@@ -171,16 +170,7 @@ abstract class AstSerializeTestMixin
 }
 
 @reflectiveTest
-class ResynthesizeAstSpecTest extends _ResynthesizeAstTest {
-  @override
-  bool get isStrongMode => false;
-}
-
-@reflectiveTest
 class ResynthesizeAstStrongTest extends _ResynthesizeAstTest {
-  @override
-  bool get isStrongMode => true;
-
   @failingTest // See dartbug.com/32290
   test_const_constructor_inferred_args() =>
       super.test_const_constructor_inferred_args();
@@ -273,13 +263,9 @@ abstract class _ResynthesizeAstTest extends ResynthesizeTest
   @override
   AnalysisOptionsImpl createOptions() {
     if (isStrongMode) {
-      return super.createOptions()
-        ..previewDart2 = true
-        ..strongMode = true;
+      return super.createOptions()..previewDart2 = true;
     } else {
-      return super.createOptions()
-        ..previewDart2 = false
-        ..strongMode = false;
+      return super.createOptions()..previewDart2 = false;
     }
   }